+++ src/grid/local_grid.rs
@@ -1,21 +1,24 @@
-//! Parallel grid
+//! Serial grids
 #[cfg(feature = "serde")]
 use crate::traits::{ConvertToSerializable, RONImportParallel};
 use crate::{
-    traits::{Entity, Grid, ParallelGrid as ParallelGridTrait},
+    traits::{Entity, Grid},
     types::Ownership,
 };
-use mpi::traits::Communicator;
 
-/// Parallel grid entity
+mod single_element;
+pub use single_element::{SingleElementGrid, SingleElementGridBorrowed, SingleElementGridBuilder};
+pub(crate) use single_element::{SingleElementGridEntity, SingleElementGridEntityBorrowed};
+
+/// Grid entity
 #[derive(Debug)]
-pub struct ParallelGridEntity<E: Entity> {
+pub struct GridEntity<E: Entity> {
     serial_entity: E,
     ownership: Ownership,
     global_index: usize,
 }
 
-impl<E: Entity> ParallelGridEntity<E> {
+impl<E: Entity> GridEntity<E> {
     /// Create new
     pub fn new(serial_entity: E, ownership: &[Ownership], global_indices: &[usize]) -> Self {
         let index = serial_entity.local_index();
@@ -26,7 +29,7 @@ impl<E: Entity> ParallelGridEntity<E> {
         }
     }
 }
-impl<E: Entity> Entity for ParallelGridEntity<E> {
+impl<E: Entity> Entity for GridEntity<E> {
     type T = E::T;
     type EntityDescriptor = E::EntityDescriptor;
     type Topology<'a>
@@ -60,15 +63,15 @@ impl<E: Entity> Entity for ParallelGridEntity<E> {
     }
 }
 
-/// Parallel grid entity iterator
+/// Grid entity iterator
 #[derive(Debug)]
-pub struct ParallelGridEntityIter<'a, E: Entity, EntityIter: Iterator<Item = E>> {
+pub struct GridEntityIter<'a, E: Entity, EntityIter: Iterator<Item = E>> {
     iter: EntityIter,
     ownership: &'a [Ownership],
     global_indices: &'a [usize],
 }
 
-impl<'a, E: Entity, EntityIter: Iterator<Item = E>> ParallelGridEntityIter<'a, E, EntityIter> {
+impl<'a, E: Entity, EntityIter: Iterator<Item = E>> GridEntityIter<'a, E, EntityIter> {
     /// Create new
     pub fn new(iter: EntityIter, ownership: &'a [Ownership], global_indices: &'a [usize]) -> Self {
         Self {
@@ -78,14 +81,12 @@ impl<'a, E: Entity, EntityIter: Iterator<Item = E>> ParallelGridEntityIter<'a, E
         }
     }
 }
-impl<E: Entity, EntityIter: Iterator<Item = E>> Iterator
-    for ParallelGridEntityIter<'_, E, EntityIter>
-{
-    type Item = ParallelGridEntity<E>;
+impl<E: Entity, EntityIter: Iterator<Item = E>> Iterator for GridEntityIter<'_, E, EntityIter> {
+    type Item = GridEntity<E>;
 
-    fn next(&mut self) -> Option<ParallelGridEntity<E>> {
+    fn next(&mut self) -> Option<GridEntity<E>> {
         let entity = self.iter.next();
-        entity.map(|e| ParallelGridEntity::new(e, self.ownership, self.global_indices))
+        entity.map(|e| GridEntity::new(e, self.ownership, self.global_indices))
     }
 }
 
@@ -149,7 +150,7 @@ impl<G: Grid + Sync> LocalGrid<G> {
 impl<G: Grid + Sync> Grid for LocalGrid<G> {
     type T = G::T;
     type Entity<'a>
-        = ParallelGridEntity<G::Entity<'a>>
+        = GridEntity<G::Entity<'a>>
     where
         Self: 'a;
     type GeometryMap<'a>
@@ -158,7 +159,7 @@ impl<G: Grid + Sync> Grid for LocalGrid<G> {
         Self: 'a;
     type EntityDescriptor = G::EntityDescriptor;
     type EntityIter<'a>
-        = ParallelGridEntityIter<'a, G::Entity<'a>, G::EntityIter<'a>>
+        = GridEntityIter<'a, G::Entity<'a>, G::EntityIter<'a>>
     where
         Self: 'a;
 
@@ -172,7 +173,7 @@ impl<G: Grid + Sync> Grid for LocalGrid<G> {
     fn entity(&self, dim: usize, serial_index: usize) -> Option<Self::Entity<'_>> {
         self.serial_grid
             .entity(dim, serial_index)
-            .map(|e| ParallelGridEntity::new(e, &self.ownership[dim], &self.global_indices[dim]))
+            .map(|e| GridEntity::new(e, &self.ownership[dim], &self.global_indices[dim]))
     }
 
     fn entity_types(&self, dim: usize) -> &[Self::EntityDescriptor] {
@@ -184,7 +185,7 @@ impl<G: Grid + Sync> Grid for LocalGrid<G> {
     }
 
     fn entity_iter(&self, dim: usize) -> Self::EntityIter<'_> {
-        ParallelGridEntityIter::new(
+        GridEntityIter::new(
             self.serial_grid.entity_iter(dim),
             &self.ownership[dim],
             &self.global_indices[dim],
@@ -194,7 +195,7 @@ impl<G: Grid + Sync> Grid for LocalGrid<G> {
     fn entity_from_id(&self, dim: usize, id: usize) -> Option<Self::Entity<'_>> {
         self.serial_grid
             .entity_from_id(dim, id)
-            .map(|e| ParallelGridEntity::new(e, &self.ownership[dim], &self.global_indices[dim]))
+            .map(|e| GridEntity::new(e, &self.ownership[dim], &self.global_indices[dim]))
     }
 
     fn geometry_map(
@@ -203,103 +204,5 @@ impl<G: Grid + Sync> Grid for LocalGrid<G> {
         points: &[Self::T],
     ) -> Self::GeometryMap<'_> {
         self.serial_grid.geometry_map(entity_type, points)
-    }
-}
-
-/// Parallel grid
-#[derive(Debug)]
-pub struct ParallelGrid<'a, C: Communicator, G: Grid + Sync> {
-    comm: &'a C,
-    local_grid: LocalGrid<G>,
-}
-
-impl<'a, C: Communicator, G: Grid + Sync> ParallelGrid<'a, C, G> {
-    /// Create new
-    pub fn new(
-        comm: &'a C,
-        serial_grid: G,
-        ownership: Vec<Vec<Ownership>>,
-        global_indices: Vec<Vec<usize>>,
-    ) -> Self {
-        Self {
-            comm,
-            local_grid: LocalGrid::new(serial_grid, ownership, global_indices),
-        }
-    }
-}
-
-#[cfg(feature = "serde")]
-impl<'a, C: Communicator + 'a, G: Grid + Sync + ConvertToSerializable> RONImportParallel<'a, C>
-    for ParallelGrid<'a, C, G>
-where
-    for<'de2> <G as ConvertToSerializable>::SerializableType: serde::Deserialize<'de2>,
-    Self: 'a,
-{
-    fn create_from_ron_info(comm: &'a C, local_grid: LocalGrid<G>) -> Self {
-        Self { comm, local_grid }
-    }
-}
-
-impl<C: Communicator, G: Grid + Sync> ParallelGridTrait<C> for ParallelGrid<'_, C, G> {
-    type LocalGrid<'a>
-        = LocalGrid<G>
-    where
-        Self: 'a;
-    fn comm(&self) -> &C {
-        self.comm
-    }
-    fn local_grid(&self) -> &LocalGrid<G> {
-        &self.local_grid
-    }
-}
-impl<C: Communicator, G: Grid + Sync> Grid for ParallelGrid<'_, C, G> {
-    type T = G::T;
-    type Entity<'a>
-        = <LocalGrid<G> as Grid>::Entity<'a>
-    where
-        Self: 'a;
-    type GeometryMap<'a>
-        = <LocalGrid<G> as Grid>::GeometryMap<'a>
-    where
-        Self: 'a;
-    type EntityDescriptor = <LocalGrid<G> as Grid>::EntityDescriptor;
-    type EntityIter<'a>
-        = <LocalGrid<G> as Grid>::EntityIter<'a>
-    where
-        Self: 'a;
-
-    fn geometry_dim(&self) -> usize {
-        self.local_grid.geometry_dim()
-    }
-    fn topology_dim(&self) -> usize {
-        self.local_grid.topology_dim()
-    }
-
-    fn entity(&self, dim: usize, local_index: usize) -> Option<Self::Entity<'_>> {
-        self.local_grid.entity(dim, local_index)
-    }
-
-    fn entity_types(&self, dim: usize) -> &[Self::EntityDescriptor] {
-        self.local_grid.entity_types(dim)
-    }
-
-    fn entity_count(&self, entity_type: Self::EntityDescriptor) -> usize {
-        self.local_grid.entity_count(entity_type)
-    }
-
-    fn entity_iter(&self, dim: usize) -> Self::EntityIter<'_> {
-        self.local_grid.entity_iter(dim)
-    }
-
-    fn entity_from_id(&self, dim: usize, id: usize) -> Option<Self::Entity<'_>> {
-        self.local_grid.entity_from_id(dim, id)
-    }
-
-    fn geometry_map(
-        &self,
-        entity_type: Self::EntityDescriptor,
-        points: &[Self::T],
-    ) -> Self::GeometryMap<'_> {
-        self.local_grid.geometry_map(entity_type, points)
     }
 }
