+++ src/grid/parallel.rs
@@ -1,5 +1,4 @@
 //! MPI parallel grids
-mod builder;
 mod grid;
 
 pub use grid::ParallelGrid;
