--- src/grid/serial.rs
@@ -1,4 +1,0 @@
-//! Serial grids
-mod single_element;
-pub use single_element::{SingleElementGrid, SingleElementGridBorrowed, SingleElementGridBuilder};
-pub(crate) use single_element::{SingleElementGridEntity, SingleElementGridEntityBorrowed};
