--- src/topology/serial.rs
@@ -1,6 +1,0 @@
-//! Serial topology
-pub(crate) mod single_type;
-pub use single_type::{
-    SingleTypeEntityTopology, SingleTypeEntityTopologyBorrowed, SingleTypeTopology,
-    SingleTypeTopologyBorrowed,
-};
