+++ src/topology/single_type.rs
